requirements.txt
docker
PyYAML
requests